requirements.txt
pytest==6.0.1
boto3 >= 1.9.160,<2.0.0
requests==2.23.0
pycryptodome >= 3.9.7
moto>=1.3.14
